rap-web/server-modules/server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.4
grpcio==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
nuitka==2.8.1
ordered-set==4.1.0
protobuf==6.33.0
psutil==7.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
zstandard==0.25.0
websockets==10.4 